requirements.txt
keras_nightly==2.5.0.dev2021032900
tensorflow==2.5.0
opencv_python==4.5.2.52
matplotlib==3.4.2
numpy==1.19.5
pandas==1.2.4
ipython==7.24.1
keras==2.4.3
Pillow==8.2.0
